Split itemstore implementation into modules
